requirements.txt
aiogram~=3.6.0
betterlogging~=1.0.0
python-dotenv~=1.0.1
aiohttp~=3.9.5
aiosqlite~=0.20.0
openpyxl~=3.1.2
qrcode~=7.4.2
pytonconnect~=0.3.1
pytoniq-core==0.1.36